+++ frontend/src/components/ContactInfo/ContactInfoEditor/index.css
@@ -1,0 +1,145 @@
+.question-header-row {
+  margin-top: 15px;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  width: 100%;
+  height: 3em;
+  align-items: center;
+}
+
+.question-number-container {
+  display: flex;
+  margin-left: 2%;
+  margin-right: 5%;
+  height: 80%;
+  width: 7%;
+  background: #d9d9d9;
+  border-radius: 15px;
+  align-items: center;
+  justify-content: center;
+}
+
+.question-number-text {
+  font-size: 16px;
+  text-align: center;
+  color: #666666;
+}
+
+.question-name-input {
+  background-color: #d9d9d9;
+  border: none;
+  border-radius: 15px;
+  height: 80%;
+  width: 55%;
+  font-size: 16px;
+  padding-left: 2%;
+  margin-right: 25%;
+}
+.question-close-button-container {
+  border-radius: 20px;
+  height: 30px;
+  width: 30px;
+  display: flex;
+  justify-content: center;
+  /* align-items: center; */
+}
+.question-close-button-container:hover {
+  background-color: rgba(50, 50, 50, 0.1);
+}
+.question-cancel-button {
+  font-size: 18px;
+  /* text-align: center;  */
+  background: transparent;
+  border: none;
+  align-self: right;
+}
+
+.single-select-question-builder-container {
+  position: relative;
+  margin-top: 15px;
+  padding: 15px;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
+  transition: 0.3s;
+  width: 70%;
+  border-radius: 5px;
+  align-items: left;
+}
+
+.contact-info-container {
+  margin-top: 2%;
+  margin-bottom: 2%;
+  align-items: center;
+  display: flex;
+  width: 100%;
+  padding-left: 5px;
+  padding-right: 5px;
+}
+
+.contact-info-container-inner {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.contact-info-row {
+  height: 3em;
+  display: flex;
+  height: 15%;
+  align-items: center;
+}
+
+.contact-info-field-label {
+  font-weight: 500;
+  color: #666666;
+  font-size: 16px;
+  margin-right: 2%;
+  margin-bottom: 2%;
+  margin-top: 1%;
+  width: 12%;
+  text-align: left;
+}
+
+.contact-info-user-input {
+  border: 1px solid #666666;
+  border-radius: 10px;
+  padding-left: 2%;
+  height: 2.5em;
+  margin-left: 2%;
+  width: 45%;
+}
+
+.contact-info-user-input:focus {
+  outline: none;
+}
+
+.question-footer-row {
+  margin: 15px;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  width: 100%;
+  height: 3em;
+  align-items: center;
+}
+
+.question-logic-added-sign {
+  font-size: 1em;
+  font-weight: 400;
+  margin-left: 2%;
+  display: flex;
+  width: 15%;
+  border: 2px solid #aeaeae;
+  color: #aeaeae;
+  border-radius: 15px;
+  height: 80%;
+  align-items: center;
+  justify-content: center;
+}
+
+.question-required-checkbox {
+  margin-right: 2%;
+  color: #aeaeae;
+}
