+++ frontend/src/components/Divider/index.css
@@ -2,6 +2,5 @@
     width: 100px;
     height: 3px;
     border-radius: 10px;
-    margin-bottom: 40px;
     margin-top: 40px;
 }
