+++ frontend/src/components/Dropdown/index.css
@@ -1,13 +1,3 @@
-.dropdownContainer {
-  width: 50%;
-}
-.label {
-  font-size: 24px;
-  font-weight: 600;
-  color: #666666;
-  margin-bottom: 20px;
-  text-align: left;
-}
 .select{
   padding: 10px;
   border-radius: 10px;
