+++ frontend/src/components/index.css
@@ -101,3 +101,16 @@
     align-items: center;
 }
 
+
+/* Customer Form Components */
+.GlobalCustomerQuestionContainer {
+    width: 50%;
+}
+
+.GlobalQuestionTitle {
+    font-size: 24px;
+    font-weight: 600;
+    color: #666666;
+    margin-bottom: 20px;
+    text-align: left;
+}
